main.py new command
I was asked to make  a  jail command. So I did! The command adds the jail role to the user that  was given with the tag and removes the @everyone  role that stops him from talking to all other channels  upon the jail one it gives  a message that the user has been jailed and that's the end of our command

ATTENTION : This command hasn't been tested yet so it's possible for the command to don't work please test it and give me feedback

I hope the command works:wlm723